feat(react): add processMultipartResponse utility for incremental delivery
