+++ packages/react/src/index.ts
@@ -2,5 +2,6 @@ export { Deferred } from "./Deferred";
 export { clientLoaderQuery, getClientLoaderQuery } from "./client-loader-query";
 export { RemixRelayProvider } from "./deferred-query-context";
 export { getCachedResponse } from "./get-cached-response";
+export { processMultipartResponse } from "./incremental-response";
 export { metaQuery } from "./meta-query";
 export { useLoaderQuery, useRouteLoaderQuery } from "./useLoaderQuery";
